+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.1] - 2025-03-04
+### Added
+- `ContainsMatcher`, `StartsWithMatcher`, `EndsWithMatcher`
+### Changed
+- `AlternateMatcher`, `CaseInsensitiveMatcher`, `PatternMatcher`, `SimpleMatcher`, `WildcardMatcher`: added `equals()`
+  and `hashCode()`
+
 ## [1.0] - 2025-03-02
 ### Added
 - all files: initial versions